requirements.txt
# GUI Dependencies
Kivy~=2.3.0
kivymd~=1.2.0
pillow~=10.2.0

# External Communication
flask~=3.0.3
requests~=2.31.0
pysmb~=1.2.9.1
aiohttp
smbprotocol
mysql-connector-python~=8.4.0

# Face Module Dependencies
opencv-python~=4.9.0.80
deepface~=0.0.91

#Speech Module Dependencies
speechrecognition
sounddevice~=0.4.6
gtts~=2.5.1
pygame
pyaudio~=0.2.13
whisper~=1.1.10

# Chatbot Module Dependencies
nltk~=3.8.1
keras~=2.12
tensorflow~=2.16.1
tf-keras~=2.16.0
numpy~=1.26.4
matplotlib~=3.9.0

# GPIO PINS
RPi.GPIO
fake-rpigpio


playsound~=1.3.0
pydub~=0.25.1
seaborn~=0.13.2
pandas~=2.2.2
tf_keras~=2.16.0
torch~=2.3.0
onnxruntime~=1.16.0
scipy~=1.13.1
websocket-client~=1.8.0
websockets~=12.0
ctranslate2~=4.3.0
tokenizers~=0.15.2
kaldialign~=0.9.1
soundfile~=0.12.1
jsonify~=0.5
python-dotenv~=1.0.1